requirements.txt
click==8.0.1
Pillow==8.2.0
requests==2.25.1
https://github.com/turicas/rows/archive/develop.zip
tqdm==4.61.0
color-extraction==0.1a4
extcolors==1.0.0
wcag-contrast-ratio==0.9
opencv-python==4.5.2.54
